refactor: move out unrelated code
