+++ src/query/stream_schema_provider.rs
@@ -442,7 +442,7 @@ impl TableProvider for StandardTableProvider {
 
         if include_now(filters, &time_partition) {
             if let Some(staging) = STAGING.get_stream(&self.stream) {
-                let records =  staging.recordbatches_cloned(&self.schema);
+                let records = staging.recordbatches_cloned(&self.schema);
                 let reversed_mem_table = reversed_mem_table(records, self.schema.clone())?;
 
                 let memory_exec = reversed_mem_table
